remove layout
