+++ src/app/(home)/blog/layout.tsx
