+++ src/app/(home)/blog/posts/diy-suspense-in-react.mdx
