+++ src/app/(home)/blog/posts/gh-static-deploy.md
