+++ src/app/reacjin/page.tsx
@@ -9,7 +9,7 @@ export const metadata: Metadata = {
 
 export default function ReacjinPage() {
   return (
-    <div className="absolute top-28 left-0 right-0 bottom-0">
+    <div className="absolute top-8 left-0 right-0 bottom-0">
       <ReacjinEditor />
     </div>
   );
